Refactor episodes/tests.py into a test package with one module per component
Split the 1,187-line monolithic test file into 9 focused modules under
episodes/tests/, improving navigability and enabling targeted test runs
(e.g. `manage.py test episodes.tests.test_models`).
